Fix integer formatting on status/shop screens, safe sizeof counts, and bank storage

+++ src/core/game.c
@@ -14,10 +14,10 @@
 void game_render_reset(Game *g)
 {
     uint8_t *p;
-    uint8_t n;
+    uint16_t n;
     if (!g) return;
     p = (uint8_t *)&g->render_cache;
-    n = (uint8_t)sizeof(RenderCache);
+    n = sizeof(RenderCache);
     while (n--) *p++ = 0xFF;
     g->render_cache.valid = false;
     g->render_cache.prev_dialogue_active = false;
